Rename internal header constant_time.h to constant_time_internal.h
