+++ library/bignum.c
@@ -41,7 +41,7 @@
 #include "bn_mul.h"
 #include "mbedtls/platform_util.h"
 #include "mbedtls/error.h"
-#include "constant_time.h"
+#include "constant_time_internal.h"
 
 #include <string.h>
