+++ library/constant_time.c
@@ -23,7 +23,7 @@
  */
 
 #include "common.h"
-#include "constant_time.h"
+#include "constant_time_internal.h"
 #include "mbedtls/constant_time.h"
 #include "mbedtls/error.h"
 #include "mbedtls/platform_util.h"
