+++ library/ssl_srv.c
@@ -34,7 +34,7 @@
 #include "mbedtls/debug.h"
 #include "mbedtls/error.h"
 #include "mbedtls/platform_util.h"
-#include "constant_time.h"
+#include "constant_time_internal.h"
 #include "mbedtls/constant_time.h"
 
 #include <string.h>
